changed to manage state in view model
